MumhaiDJ/requirements.txt
asgiref==3.2.3
astroid==2.3.3
bson==0.5.8
colorama==0.4.3
dataclasses==0.6
Django==3.0.7
djongo==1.3.1
dnspython==1.16.0
et-xmlfile==1.0.1
IMDbPY==6.8
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
mongoengine==0.19.1
numpy==1.19.1
openpyxl==3.0.4
pandas==1.0.5
Pillow==7.1.1
psycopg2==2.8.4
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.18
sqlparse==0.2.4
toml==0.10.1
winkerberos==0.7.0
wrapt==1.11.2
